requirements.txt
# =============================================================================
# AI Brand Visibility Intelligence Engine — Python dependencies
# =============================================================================

# --- Core ---
psycopg2-binary>=2.9,<3.0
openai>=1.30,<2.0
anthropic>=0.30,<1.0

# --- Data science ---
numpy>=1.26,<3.0
scikit-learn>=1.4,<2.0
hdbscan>=0.8.33,<1.0

# --- Dashboard ---
streamlit>=1.35,<2.0
plotly>=5.20,<6.0
pandas>=2.2,<3.0

# --- Testing ---
pytest>=8.0,<9.0
pytest-cov>=5.0,<6.0

# --- Utilities ---
python-dotenv>=1.0,<2.0
